pass storeId along shopping flow to checkout page

+++ app/src/main/java/edu/rentals/frontend/ShoppingCheckoutRental.java
@@ -8,12 +8,12 @@ import java.sql.Timestamp;
 import java.util.List;
 
 public class ShoppingCheckoutRental {
-    private int storeId;
+    private long storeId;
     private String userId;
     private List<Equipment> equipment;
     private Timestamp startTimestamp, endTimestamp;
 
-    public ShoppingCheckoutRental(int storeId, String userId, Timestamp startTimestamp, Timestamp endTimestamp, List<Equipment> equipment) {
+    public ShoppingCheckoutRental(long storeId, String userId, Timestamp startTimestamp, Timestamp endTimestamp, List<Equipment> equipment) {
         this.storeId = storeId;
         this.userId = userId;
         this.equipment = equipment;
